Refractors TooltipIsLocked to its own file
